refactor(mirror): split score/tests.rs (1313) into themed sub-files (#121)
VibeGuard U-16 sets 800 lines as the hard ceiling. score/tests.rs was at
1313, the only file in the repo over that limit.

Split by themed test groups, all helpers consolidated in tests/mod.rs:
- tests/mod.rs:         291 (shared helpers + sub-module declarations)
- tests/baseline.rs:    367 (8 personal_baseline / apply_baseline tests)
- tests/compute.rs:     187 (6 knowledge_rate / friction_density / layer tests)
- tests/persistence.rs: 190 (5 persist / load / rotate / concurrent tests)
- tests/signal.rs:      132 (5 signal conversion / dreyfus / tension tests)
- tests/paths.rs:        86 (6 growth_tracker_path + filter_since tests)
- tests/streak.rs:       76 (7 streak count + milestone + display tests)

Every file is now well under 400 lines.

No production code changed. No test logic, assertion text, or test ordering
within each theme changed. `cargo test -p mirror`: 88 passed, 0 failed.

Closes #121
